requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.0.8
httpx==0.27.0
openpyxl==3.1.5
python-dotenv==1.0.1
